Allow test logging setup to be reused
